requirements.txt
Django>=3.2.4<4.0
django-filter==2.4.0
django-guardian==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn>=20.1.0
